update uploadplan regex to fix today's bug; add fallback default text
There was another "Upload-Plan am" in the string which caused the regex to select the wrong paragraph. It now makes sure to use only "Uploadplan am digit digit dot" with a new line in front of it and a text length of 100 or more for the uploadplan. This is backwards compatible (as always) and should avoid future bugs like this.
It also falls back to a default text when no regex match was found